#1691 Last attempt at current version. Going to try somehting else as this isn't working

+++ source/tests/SGM/AMScalerAnalysisTest.h
@@ -14,7 +14,7 @@ signals:
 public slots:
 protected:
 
-	void AnalyseVector(const QVector<int>& counts);
+	void AnalyseVector(const QVector<double>& counts);
 
 };
